feat: implements api route, service and controller to /packs

+++ backend/src/database/migrations/20230514014938-create-products.js
@@ -4,18 +4,18 @@ module.exports = {
   async up (queryInterface, Sequelize) {
     await queryInterface.createTable('products', {
       code: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT,
         primaryKey: true,
         allowNull: false,
       },
       name: { type: Sequelize.STRING, allowNull: false },
       costPrice: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false,
         field: 'cost_price',
       },
       salesPrice: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false,
         field: 'sales_price',
       },
